refactor(dfpt): split DFPT_PW init/build_occ_kq and run dispatchers into dedicated TUs
Moves DFPT_PW::init / DFPT_PW::Impl::build_occ_kq out of dfpt_pw.cpp into
dfpt_pw_init.cpp and DFPT_PW::run into dfpt_pw_run.cpp, extracting helper
member functions so every moved function keeps cyclomatic complexity
under the coding-rule 10 budget. init() is now a three-stage flow
(check_metallic_occ / setup_q_list / init_submodules), build_occ_kq is
split into (match_commensurate_kq / copy_occ_state_ball) and the run()
scheduler is split into (run_q0_pre / run_displacement_irrep_pass /
run_q0_post / run_assemble). Keeps the public DFPT_PW API 100% compatible
and updates all three CMakeLists entries deterministically.

Verification commands:
  cmake --build build -j 30                         -> PASS (no errors,
                                                       module_dfpt links
                                                       cleanly)
  cmake --build build_max_para_test -j 30 \
    --target MODULE_DFPT_pw_run_test                -> PASS
  cmake --build build_max_para_test -j 30 \
    --target MODULE_DFPT_{pw_data_test,pert_serial,\
    phon_serial,q0_serial,rho_serial}               -> PASS all
  OMP_NUM_THREADS=1 ctest --test-dir build_max_para_test \
    -R MODULE_DFPT -V                               -> 8/8 tests passed
                                                       (5.59 s elapsed)
  python3 tools/03_code_analysis/agent_governance_check.py \
    --staged                                        -> no findings
