+++ source/source_pw/module_dfpt/CMakeLists.txt
@@ -2,6 +2,8 @@ set(MODULE_NAME module_dfpt)
 
 set(SOURCES
     dfpt_pw.cpp
+    dfpt_pw_init.cpp
+    dfpt_pw_run.cpp
     dfpt_pw_data.cpp
     dfpt_kq_basis.cpp
     dfpt_pert.cpp
