+++ source/source_pw/module_dfpt/dfpt_pw.cpp
@@ -51,160 +51,6 @@ DFPT_PW::~DFPT_PW()
 {
 }
 
-void DFPT_PW::init(const DFPT_PW_InitContext& ctx)
-{
-    ModuleBase::TITLE("DFPT_PW", "init");
-    ModuleBase::timer::start("DFPT_PW", "init");
-    // ctx.{psi,wg,veff_r,eig} are valid non-null pointers even in skeleton
-    // mode: the empty Psi / matrix objects still carry the queryable nk /
-    // nbands / nrxx shape fields consumed by the init helpers; ucell is
-    // always a non-null per the public wrapper (it takes a reference).
-    UnitCell* const ucell = ctx.ucell;
-    const psi::Psi<std::complex<double>>* const psi = ctx.psi;
-    ModulePW::PW_Basis* const pw_rho = ctx.pw_rho;
-    ModulePW::PW_Basis_K* const pw_wfc = ctx.pw_wfc;
-    Structure_Factor* const sf = ctx.sf;
-    const std::vector<double>* const veff_r = ctx.veff_r;
-    const ModuleBase::matrix* const wg = ctx.wg;
-    const ModuleBase::matrix* const eig = ctx.eig;
-    const XC_First_Order* const xc = ctx.xc;
-    const double nelec = ctx.nelec;
-    const double ecutwfc = ctx.ecutwfc;
-    const Plus_U_Base* const dftu = ctx.dftu;
-
-    pimpl_->ucell_ = ucell;
-    pimpl_->gs_psi_ = *psi;
-    pimpl_->pw_rho_ = pw_rho;
-    pimpl_->pw_wfc_ = pw_wfc;
-    pimpl_->sf_ = sf;
-    pimpl_->veff_r_ = *veff_r;
-    pimpl_->wg_ = *wg;
-    pimpl_->eig_ = *eig;
-
-    // Metallic-sampling guard: the Sternheimer/projector flow treats every
-    // band as either fully occupied or empty and carries no d(mu)/dtau
-    // response, so a sampling whose smearing Fermi level cuts a band (wg
-    // strictly between 0 and the full reference) yields force constants
-    // wrong at the 100% level while still converging cleanly. Reject it
-    // explicitly (C4 defers metallic DFPT); negligible gauss tails
-    // (relative weight < 1e-3) are tolerated as the insulator limit.
-    const double frac_weight_tol = 1.0e-3; ///< empirical parameter: relative band weight treated as metallic
-    for (int ik = 0; ik < wg->nr; ++ik)
-    {
-        const double wref = (*wg)(ik, 0);
-        if (wref <= 0.0)
-        {
-            continue;
-        }
-        for (int ib = 0; ib < wg->nc; ++ib)
-        {
-            const double rel = (*wg)(ik, ib) / wref;
-            if (rel > frac_weight_tol && rel < 1.0 - frac_weight_tol)
-            {
-                std::stringstream msg;
-                msg << "fractional band occupation at (ik=" << ik << ", ib=" << ib << ", wg=" << (*wg)(ik, ib)
-                    << "): metallic DFPT (smearing occupations crossing the"
-                       " Fermi level) is not supported; reduce smearing sigma"
-                       " or use an insulating k sampling.";
-                ModuleBase::WARNING_QUIT("DFPT_PW::init", msg.str());
-            }
-        }
-    }
-    pimpl_->xc_ = xc;
-    pimpl_->nelec_ = nelec;
-    pimpl_->ecutwfc_ = ecutwfc;
-    pimpl_->dftu_ = dftu;
-
-    // DFT+U guard: the ground state now supports PW-basis DFT+U and wires a
-    // provider when dft_plus_u is enabled, but every DFPT U hook
-    // (DFPT_Rho::cal_docc, DFPT_Pert::build_dv_u, DFPT_Q0 born/docc
-    // contractions, DFPT_Phon::dftu_onsite) is a no-op reservation (U0).
-    // Running anyway would converge cleanly while silently dropping the
-    // whole first-order U response, so reject explicitly until U1 lands
-    // (same fail-loud pattern as the metallic-sampling guard above).
-    if (dftu != nullptr)
-    {
-        ModuleBase::WARNING_QUIT("DFPT_PW::init",
-                                 "DFT+U with DFPT is not supported yet: the "
-                                 "first-order U response is not implemented "
-                                 "(U0 reservation); rerun with dft_plus_u 0.");
-    }
-
-    // q points: an explicit q list file overrides the Monkhorst-Pack mesh
-    if (!pimpl_->qfile_.empty())
-    {
-        pimpl_->qlist_.read_from_file(pimpl_->qfile_, *ucell);
-        if (pimpl_->qlist_.get_nq() == 0)
-        {
-            ModuleBase::WARNING_QUIT("DFPT_PW::init", "failed to read the DFPT q-point file: " + pimpl_->qfile_);
-        }
-    }
-    else
-    {
-        std::vector<int> mp_grid = {pimpl_->nqx_, pimpl_->nqy_, pimpl_->nqz_};
-        pimpl_->qlist_.generate_mesh(*ucell, ucell->symm, mp_grid, true);
-    }
-
-    const int nq = pimpl_->qlist_.get_nq();
-    const int nk = psi->get_nk();
-    const int nbands = psi->get_nbands();
-    const int npw_max = psi->get_current_ngk();
-    const int nrxx = (pw_rho != nullptr) ? pw_rho->nrxx : 0;
-    const int nspin = 1;
-    const int nat = ucell->nat;
-
-    if (pw_rho != nullptr && pw_wfc != nullptr && sf != nullptr)
-    {
-        pimpl_->pert_.init(*ucell, pw_rho, pw_wfc, *sf);
-        // plain-mixing coefficient: the response Jacobian has strongly
-        // negative eigenvalues concentrated on the smallest-G shells (the
-        // Coulomb stiffness 4pi/G^2; measured lambda ~ -2.2 on {111}/{200}
-        // for the diamond smoke case), so the coefficient must stay below
-        // 2 / (1 + |lambda_min|); the INPUT default 0.4 keeps margin up to
-        // |lambda| ~ 3; the alternative is mix_type = "kerker", the screen
-        // f_g = |G+q|^2 / (|G+q|^2 + a^2) in 1/lat0^2 units (a^2 via
-        // DFPT_KERKER_A2), which stabilizes those shells at beta up to 1;
-        // the env knobs are design-phase calibration aids
-        double mix_beta = pimpl_->mix_beta_;
-        if (const char* env_beta = getenv("DFPT_MIX_BETA"))
-        {
-            const double parsed = atof(env_beta);
-            if (parsed > 0.0 && parsed <= 1.0)
-            {
-                mix_beta = parsed;
-            }
-        }
-        std::string mix_type = "plain";
-        if (const char* env_type = getenv("DFPT_MIX_TYPE"))
-        {
-            const std::string parsed = env_type;
-            if (parsed == "plain" || parsed == "kerker")
-            {
-                mix_type = parsed;
-            }
-        }
-        double kerker_a2 = 1.0;
-        if (const char* env_a2 = getenv("DFPT_KERKER_A2"))
-        {
-            const double parsed = atof(env_a2);
-            if (parsed > 0.0)
-            {
-                kerker_a2 = parsed;
-            }
-        }
-        pimpl_->rho_.init({nspin, nrxx, pw_rho, pw_wfc, ucell->G, mix_type, mix_beta, kerker_a2});
-        pimpl_->phon_.init(*ucell, pw_rho, &pimpl_->pert_);
-        pimpl_->q0_.init(*ucell, pw_rho, pw_wfc, &pimpl_->pert_);
-        pimpl_->hamilt_.reset(new DFPT_HamiltShift(*ucell, pw_rho, pw_wfc, *veff_r, &pimpl_->pert_));
-    }
-    else
-    {
-        pimpl_->phon_.init(*ucell, nullptr, nullptr);
-    }
-    pimpl_->data_.init(&pimpl_->qlist_, nk, nbands, npw_max, nrxx, nspin, nat, dftu);
-    ModuleBase::timer::end("DFPT_PW", "init");
-}
-
 bool DFPT_PW::get_with_u() const
 {
     return pimpl_->data_.with_u();
@@ -213,113 +59,6 @@ bool DFPT_PW::get_with_u() const
 bool DFPT_PW::get_u_active() const
 {
     return pimpl_->data_.u_active();
-}
-
-void DFPT_PW::Impl::build_occ_kq(int q_idx)
-{
-    ModuleBase::TITLE("DFPT_PW", "build_occ_kq");
-    ModuleBase::timer::start("DFPT_PW", "build_occ_kq");
-    const int nk = pw_wfc_->nks;
-    occ_kq_.assign(nk, std::vector<std::vector<std::complex<double>>>());
-    ikq_of_k_.assign(nk, -1);
-    const ModuleBase::Vector3<double> q_frac = data_.get_qvec(q_idx);
-    const ModuleBase::Vector3<double> q_cart = q_frac * ucell_->G;
-    const double kmatch_tol = 1.0e-6; ///< empirical parameter: folded fractional k-list match tolerance
-    for (int ik = 0; ik < nk; ++ik)
-    {
-        // k+q folded into [0,1) direct coordinates must be a ground-state k
-        // point (DFPT q meshes are commensurate with the k mesh)
-        const ModuleBase::Vector3<double> target = pw_wfc_->kvec_d[ik] + q_frac;
-        int ikq = -1;
-        for (int j = 0; j < nk; ++j)
-        {
-            const ModuleBase::Vector3<double>& kj = pw_wfc_->kvec_d[j];
-            const double rx = std::round(kj.x - target.x);
-            const double ry = std::round(kj.y - target.y);
-            const double rz = std::round(kj.z - target.z);
-            if (std::abs(kj.x - target.x - rx) < kmatch_tol && std::abs(kj.y - target.y - ry) < kmatch_tol
-                && std::abs(kj.z - target.z - rz) < kmatch_tol)
-            {
-                ikq = j;
-                break;
-            }
-        }
-        if (ikq < 0)
-        {
-            std::ostringstream oss;
-            oss << "k+q is not a point of the ground-state k list: the DFPT "
-                   "q mesh must be commensurate with the k mesh (and inside "
-                   "the first Brillouin zone). ik="
-                << ik << " k_d=(" << pw_wfc_->kvec_d[ik].x << "," << pw_wfc_->kvec_d[ik].y << ","
-                << pw_wfc_->kvec_d[ik].z << ") q_d=(" << q_frac.x << "," << q_frac.y << "," << q_frac.z << ") k+q=("
-                << target.x << "," << target.y << "," << target.z << ") nk=" << nk;
-            ModuleBase::WARNING_QUIT("DFPT_PW::build_occ_kq", oss.str());
-        }
-        ikq_of_k_[ik] = ikq;
-
-        DFPT_KQ_Basis kq;
-        kq.init(pw_wfc_, pw_rho_, q_cart, ik);
-        const int npw_kq = kq.get_npwk();
-
-        // The congruence match above may fold k+q onto a *different label*
-        // of the same physical point (e.g. k lists holding both (1/2,0,0)
-        // and (-1/2,0,0), which differ by a reciprocal lattice vector).
-        // The two balls then enumerate different G labels: a state of the
-        // ikq ball with vector G' coincides physically with the k+q-ball
-        // vector G when G' + k(ijq) == G + k(ik) + q, i.e.
-        // G' = G + dn with dn = k_d(ik) + q - k_d(ikq) integer in
-        // reciprocal-basis coordinates. Coincident FFT cells identify the
-        // same G only for dn = 0, so match through the G vectors instead.
-        const ModuleBase::Vector3<double> dn = pw_wfc_->kvec_d[ik] + q_frac - pw_wfc_->kvec_d[ikq];
-        const double dnr[3] = {std::round(dn.x), std::round(dn.y), std::round(dn.z)};
-        if (std::abs(dn.x - dnr[0]) > kmatch_tol || std::abs(dn.y - dnr[1]) > kmatch_tol
-            || std::abs(dn.z - dnr[2]) > kmatch_tol)
-        {
-            ModuleBase::WARNING_QUIT("DFPT_PW::build_occ_kq",
-                                     "k+q folds onto a k-list entry with a "
-                                     "non-integer reciprocal offset.");
-        }
-        const int dn_i[3] = {static_cast<int>(dnr[0]), static_cast<int>(dnr[1]), static_cast<int>(dnr[2])};
-        const ModuleBase::Matrix3 ginv = pw_wfc_->G.Inverse();
-        // reciprocal-basis integer triple -> per-k index of the ikq ball
-        // (pw_wfc_ is a PW_Basis_K whose gcar holds a per-k ball layout,
-        // not the parent-class global-ig layout: read it through getgcar)
-        std::map<std::vector<int>, int> jgl_of_n;
-        for (int jgl = 0; jgl < pw_wfc_->npwk[ikq]; ++jgl)
-        {
-            const ModuleBase::Vector3<double> gf = pw_wfc_->getgcar(ikq, jgl) * ginv;
-            const std::vector<int> key = {static_cast<int>(std::round(gf.x)),
-                                          static_cast<int>(std::round(gf.y)),
-                                          static_cast<int>(std::round(gf.z))};
-            jgl_of_n[key] = jgl;
-        }
-
-        const int nbands = gs_psi_.get_nbands();
-        for (int m = 0; m < nbands; ++m)
-        {
-            if (!dfpt_band_occupied(wg_, ikq, m))
-            {
-                continue; // empty at k+q: outside the P_c projector
-            }
-            std::vector<std::complex<double>> state(npw_kq, std::complex<double>(0.0, 0.0));
-            for (int igl = 0; igl < npw_kq; ++igl)
-            {
-                const ModuleBase::Vector3<double> gf = kq.get_gcar(igl) * ginv;
-                const std::vector<int> key = {static_cast<int>(std::round(gf.x)) + dn_i[0],
-                                              static_cast<int>(std::round(gf.y)) + dn_i[1],
-                                              static_cast<int>(std::round(gf.z)) + dn_i[2]};
-                const auto it = jgl_of_n.find(key);
-                if (it != jgl_of_n.end())
-                {
-                    state[igl] = gs_psi_(ikq, m, it->second);
-                }
-            }
-            occ_kq_[ik].push_back(std::move(state));
-        }
-    }
-    last_q_ = q_idx;
-    last_ik_ = -1;
-    ModuleBase::timer::end("DFPT_PW", "build_occ_kq");
 }
 
 double DFPT_PW::Impl::solve_displacement(int q_idx, int iat, int idir)
@@ -829,125 +568,6 @@ void DFPT_PW::Impl::solve_efield_resp(int q_idx)
         data_.set_dpsi_efield(a, de);
     }
     ModuleBase::timer::end("DFPT_PW", "solve_efield_resp");
-}
-
-void DFPT_PW::run()
-{
-    ModuleBase::TITLE("DFPT_PW", "run");
-    ModuleBase::timer::start("DFPT_PW", "run");
-    const int nq = pimpl_->qlist_.get_nq();
-    for (int q_idx = 0; q_idx < nq; ++q_idx)
-    {
-        // Special handling for q=0 (uniform electric field responses):
-        // The standard position operator r is ill-defined in periodic systems.
-        // Developers should NOT pass a conventional position matrix. Instead,
-        // matrix elements should be computed using the well-defined periodic
-        // commutator [Ĥ_SCF, r̂]. This is implemented in DFPT_Q0 module.
-        if (q_idx == 0 && pimpl_->data_.get_compute_q0())
-        {
-            pimpl_->q0_.compute_q0_response(pimpl_->data_);
-        }
-
-        // occupied states at k+q for every k of this q (projector of P_c);
-        // also invalidates the shifted-operator context cache
-        if (pimpl_->wired())
-        {
-            pimpl_->build_occ_kq(q_idx);
-        }
-
-        // position legs of the screened Born charges: the q = 0 Y solves
-        // need the projector just built and must land before the two-pass
-        // displacement solves below reuse the shifted-operator context
-        if (q_idx == 0 && pimpl_->data_.get_compute_q0() && pimpl_->wired())
-        {
-            pimpl_->solve_pos_resp(q_idx);
-            // SCF E-field responses of the dielectric tensor: after the
-            // bare Y legs they consume, before the displacement solves
-            // reuse the slots; the epsilon contraction runs straight after
-            // (QE solve_e -> dielec.f90 order)
-            pimpl_->solve_efield_resp(q_idx);
-            pimpl_->q0_.compute_eps(pimpl_->wg_, pimpl_->data_);
-        }
-
-        // Per-irrep self-consistent loop: the little-group irrep
-        // decomposition is a placeholder until stage A, so the single
-        // available irrep falls back to the full 3N displacement basis.
-        // Ledger semantics (B4): one outer pass solves every displacement
-        // to its own convergence (solve_displacement restarts each from a
-        // zero input density), and the pass residual is the worst final
-        // displacement residual; the pass converges when that worst is
-        // below conv_thr. An unconverged pass therefore re-runs the full
-        // solve, bounded by max_iter_ outer passes, and the residual
-        // history keeps an honest record instead of the former
-        // unconditional single-pass convergence.
-        const int nirr = pimpl_->data_.get_nirr(q_idx);
-        for (int irrep = 0; irrep < nirr; ++irrep)
-        {
-            pimpl_->data_.set_converged(q_idx, irrep, false);
-            pimpl_->data_.set_current_iter(q_idx, irrep, 0);
-            while (!pimpl_->data_.get_converged(q_idx, irrep)
-                   && pimpl_->data_.get_current_iter(q_idx, irrep) < pimpl_->max_iter_)
-            {
-                if (pimpl_->wired())
-                {
-                    const int nat = pimpl_->ucell_->nat;
-                    // two passes over the 3N displacement basis: first solve
-                    // every displacement to convergence (the 2n+1 accumulation
-                    // of displacement b needs the converged dpsi AND screened
-                    // potential of every column displacement a), then run the
-                    // 2n+1 accumulation for each
-                    double worst = 0.0;
-                    for (int iat = 0; iat < nat; ++iat)
-                    {
-                        for (int idir = 0; idir < 3; ++idir)
-                        {
-                            const double residual = pimpl_->solve_displacement(q_idx, iat, idir);
-                            worst = std::max(worst, residual);
-                        }
-                    }
-                    for (int iat = 0; iat < nat; ++iat)
-                    {
-                        for (int idir = 0; idir < 3; ++idir)
-                        {
-                            // 2n+1 accumulation of this converged displacement
-                            pimpl_->phon_
-                                .accumulate_electron(q_idx, iat, idir, pimpl_->gs_psi_, pimpl_->wg_, pimpl_->data_);
-                        }
-                    }
-                    pimpl_->data_.add_residual(q_idx, irrep, worst);
-                    pimpl_->data_.set_converged(q_idx, irrep, worst < pimpl_->data_.get_conv_thr());
-                }
-                else
-                {
-                    // design-phase skeleton: no bases wired, converge at once
-                    pimpl_->data_.add_residual(q_idx, irrep, 0.0);
-                    pimpl_->data_.set_converged(q_idx, irrep, true);
-                }
-                pimpl_->data_.set_current_iter(q_idx, irrep, pimpl_->data_.get_current_iter(q_idx, irrep) + 1);
-            }
-        }
-
-        // screened Born charges: the Gonze-Lee 2n+1 form consumes the
-        // converged (screened) dpsi of every q = 0 displacement stashed by
-        // solve_displacement, so it must run after the two-pass solves
-        // above and before the LO-TO term below consumes it
-        if (q_idx == 0 && pimpl_->data_.get_compute_q0() && pimpl_->wired())
-        {
-            pimpl_->q0_.compute_born(pimpl_->gs_psi_, pimpl_->wg_, pimpl_->eig_, pimpl_->data_);
-        }
-
-        pimpl_->phon_.assemble(q_idx, pimpl_->data_);
-        pimpl_->phon_.diagonalize(q_idx, pimpl_->data_);
-        if (q_idx == 0 && pimpl_->data_.get_loto())
-        {
-            // non-analytic LO-TO correction along the data-layer direction
-            // (default isotropic (1,1,1)/sqrt(3) for cubic crystals;
-            // set_loto_dir overrides, e.g. per irrep direction in stage A)
-            pimpl_->phon_.add_loto(pimpl_->data_.get_loto_dir(), pimpl_->data_);
-            pimpl_->phon_.diagonalize_loto(pimpl_->data_);
-        }
-    }
-    ModuleBase::timer::end("DFPT_PW", "run");
 }
 
 int DFPT_PW::get_nq() const
